Removed prevent_destroy flag from bucket
The bucket resource already has protection aganist accidental deletion (it can not be removed if any object exists within the bucket). Furthermore, a public module should not have the prevent_destroy flag set unconditionally as the module code can not be modified by the user and this results in undeletable (with terraform) resources. If needed, using a conditional (and most probably optional) flag should be defined on module input to manage the prevent_destroy flags.